error alert message

+++ StoneQuest/src/com/barelyconscious/worlds/common/shape/Vector.java
@@ -13,20 +13,20 @@ public final class Vector {
     public static final Vector LEFT = new Vector(-1, 0);
     public static final Vector RIGHT = new Vector(1, 0);
 
-    public final float x;
-    public final float y;
+    public final double x;
+    public final double y;
 
     public Vector() {
         this(0, 0);
     }
 
-    public Vector(final float x, final float y) {
+    public Vector(final double x, final double y) {
         this.x = x;
         this.y = y;
     }
 
     public Vector unitVector() {
-        final float magnitude = magnitude();
+        final double magnitude = magnitude();
         if (magnitude == 0) {
             return new Vector(0, 0);
         }
@@ -36,22 +36,22 @@ public final class Vector {
             y / magnitude);
     }
 
-    public float magnitude() {
-        float sqSum = (x * x) + (y * y);
+    public double magnitude() {
+        double sqSum = (x * x) + (y * y);
         if (sqSum == 0) {
             return 0;
         }
 
-        return (float) Math.sqrt(sqSum);
+        return (double) Math.sqrt(sqSum);
     }
 
-    public Vector multiply(final float amt) {
+    public Vector multiply(final double amt) {
         return new Vector(
             x * amt,
             y * amt);
     }
 
-    public Vector plus(final float x, final float y) {
+    public Vector plus(final double x, final double y) {
         return new Vector(
             this.x + x,
             this.y + y);
